requirements/development.txt
-r base.txt
pytest-django==4.9.0
pytest==8.3.4
black==24.10.0
flake8==7.1.1